+++ aeon/visualisation/__init__.py
@@ -25,14 +25,9 @@ from aeon.visualisation.estimator._temporal_importance_curves import (
 from aeon.visualisation.results._boxplot import plot_boxplot_median
 from aeon.visualisation.results._critical_difference import plot_critical_difference
 from aeon.visualisation.results._scatter import plot_scatter, plot_scatter_predictions
+from aeon.visualisation.series._forecasting import plot_windows
 from aeon.visualisation.series._segmentation import (
     plot_time_series_with_change_points,
     plot_time_series_with_profiles,
 )
-from aeon.visualisation.series._series import (
-    plot_correlations,
-    plot_interval,
-    plot_lags,
-    plot_series,
-    plot_windows,
-)
+from aeon.visualisation.series._series import plot_correlations, plot_lags, plot_series
