+++ aeon/visualisation/series/_series.py
@@ -1,10 +1,10 @@
 """Common timeseries plotting functionality."""
 
-__all__ = ["plot_series", "plot_correlations", "plot_windows"]
+__all__ = ["plot_series", "plot_lags", "plot_correlations"]
 __author__ = ["mloning", "RNKuhns", "Drishti Bhasin", "chillerobscuro"]
 
 import math
-from warnings import simplefilter, warn
+from warnings import warn
 
 import numpy as np
 import pandas as pd
@@ -169,6 +169,18 @@ def _plot_interval(ax, interval_df):
     )
     ax.legend()
     return ax
+
+
+def _check_colors(colors, n_series):
+    """Verify color list is correct length and contains only colors."""
+    from matplotlib.colors import is_color_like
+
+    if n_series == len(colors) and all([is_color_like(c) for c in colors]):
+        return True
+    warn(
+        "Color list must be same length as `series` and contain only matplotlib colors"
+    )
+    return False
 
 
 def plot_lags(series, lags=1, suptitle=None):
@@ -356,92 +368,3 @@ def plot_correlations(
         fig.suptitle(suptitle, size="xx-large")
 
     return fig, np.array(fig.get_axes())
-
-
-def _check_colors(colors, n_series):
-    """Verify color list is correct length and contains only colors."""
-    from matplotlib.colors import is_color_like
-
-    if n_series == len(colors) and all([is_color_like(c) for c in colors]):
-        return True
-    warn(
-        "Color list must be same length as `series` and contain only matplotlib colors"
-    )
-    return False
-
-
-def _get_windows(cv, y):
-    """Generate cv split windows, utility function."""
-    train_windows = []
-    test_windows = []
-    for train, test in cv.split(y):
-        train_windows.append(train)
-        test_windows.append(test)
-    return train_windows, test_windows
-
-
-def plot_windows(cv, y, title=""):
-    """Plot training and test windows.
-
-    Parameters
-    ----------
-    y : pd.Series
-        Time series to split
-    cv : temporal cross-validation iterator object
-        Temporal cross-validation iterator
-    title : str
-        Plot title
-    """
-    _check_soft_dependencies("matplotlib", "seaborn")
-    import matplotlib.pyplot as plt
-    import seaborn as sns
-    from matplotlib.ticker import MaxNLocator
-
-    simplefilter("ignore", category=UserWarning)
-
-    train_windows, test_windows = _get_windows(cv, y)
-
-    def get_y(length, split):
-        # Create a constant vector based on the split for y-axis."""
-        return np.ones(length) * split
-
-    n_splits = len(train_windows)
-    n_timepoints = len(y)
-    len_test = len(test_windows[0])
-
-    train_color, test_color = sns.color_palette("colorblind")[:2]
-
-    fig, ax = plt.subplots(figsize=plt.figaspect(0.3))
-
-    for i in range(n_splits):
-        train = train_windows[i]
-        test = test_windows[i]
-
-        ax.plot(
-            np.arange(n_timepoints), get_y(n_timepoints, i), marker="o", c="lightgray"
-        )
-        ax.plot(
-            train,
-            get_y(len(train), i),
-            marker="o",
-            c=train_color,
-            label="Window",
-        )
-        ax.plot(
-            test,
-            get_y(len_test, i),
-            marker="o",
-            c=test_color,
-            label="Forecasting horizon",
-        )
-    ax.invert_yaxis()
-    ax.yaxis.set_major_locator(MaxNLocator(integer=True))
-    ax.set(
-        title=title,
-        ylabel="Window number",
-        xlabel="Time",
-        xticklabels=y.index,
-    )
-    # remove duplicate labels/handles
-    handles, labels = [(leg[:2]) for leg in ax.get_legend_handles_labels()]
-    ax.legend(handles, labels)
